requirements.txt
alembic==1.7.7
aniso8601==9.0.1
apispec==5.1.1
attrs==21.2.0
awsebcli==3.20.2
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
blessed==1.19.0
botocore==1.21.65
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.2
colorama==0.4.3
cryptography==35.0.0
distlib==0.3.3
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.3.2
Flask==2.1.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mysql-connector-python==8.0.28
packaging==21.3
paramiko==2.8.0
pathspec==0.5.9
platformdirs==2.4.0
protobuf==3.20.0
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==3.0.8
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==1.4.35
termcolor==1.1.0
texttable==1.6.4
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.1.9
webargs==8.1.0
websocket-client==0.59.0
Werkzeug==2.1.1
zipp==3.8.0
